+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/pusher/DrainPusher.java
@@ -1,0 +1,31 @@
+package com.dwarfeng.rbacds.impl.handler.pusher;
+
+import com.dwarfeng.rbacds.sdk.handler.pusher.AbstractPusher;
+import org.springframework.stereotype.Component;
+
+/**
+ * 简单的丢弃掉所有信息的推送器。
+ *
+ * @author DwArFeng
+ * @since 1.8.0
+ */
+@Component
+public class DrainPusher extends AbstractPusher {
+
+    public static final String PUSHER_TYPE = "drain";
+
+    public DrainPusher() {
+        super(PUSHER_TYPE);
+    }
+
+    @Override
+    public void permissionFilterReset() {
+    }
+
+    @Override
+    public String toString() {
+        return "DrainPusher{" +
+                "pusherType='" + pusherType + '\'' +
+                '}';
+    }
+}