+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/handler/Pusher.java
@@ -1,0 +1,25 @@
+package com.dwarfeng.rbacds.sdk.handler;
+
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+
+/**
+ * 事件推送器。
+ *
+ * @author DwArFeng
+ * @since 1.8.0
+ */
+public interface Pusher {
+
+    /**
+     * 返回制造器是否支持指定的类型。
+     *
+     * @param type 指定的类型。
+     * @return 制造器是否支持指定的类型。
+     */
+    boolean supportType(String type);
+
+    /**
+     * 权限过滤重置时执行的推送操作。
+     */
+    void permissionFilterReset() throws HandlerException;
+}