+++ rbac-distributed-service-stack/src/main/java/com/dwarfeng/rbacds/stack/handler/PushHandler.java
@@ -1,0 +1,18 @@
+package com.dwarfeng.rbacds.stack.handler;
+
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import com.dwarfeng.subgrade.stack.handler.Handler;
+
+/**
+ * 推送器处理器。
+ *
+ * @author DwArFeng
+ * @since 1.8.0
+ */
+public interface PushHandler extends Handler {
+
+    /**
+     * 权限过滤重置时执行的推送操作。
+     */
+    void permissionFilterReset() throws HandlerException;
+}